[ADD]Alterações no front
Alteração e customização da sidebar;
Criação da página de Contatos;
Criação da homepage;
Adição de partículas por todo o dash.
